Optimize parsers aggregating multiple values in a list
Using Value.aggregate to append an element to a list
implicitly creates a new copy of the list every time.
When parsing a large enough number of values (e.g. 100000),
the parsing time ends up dominated by the time it takes to
copy the list.

Instead, use a single mutable list and append to it.